+++ audio/upmixer_test.go
@@ -8,12 +8,12 @@ import (
 )
 
 func TestUpmixer(t *testing.T) {
-	samples := []float64{0.1, 0.2, 0.3}
-	want := []float64{0.1, 0.1, 0.2, 0.2, 0.3, 0.3}
+	samples := []float32{0.1, 0.2, 0.3}
+	want := []float32{0.1, 0.1, 0.2, 0.2, 0.3, 0.3}
 
 	upmix := audio.NewUpmixer(audio.NewReader(samples), 2)
 
-	got := make([]float64, len(samples)*2)
+	got := make([]float32, len(samples)*2)
 	_, err := upmix.ReadSamples(got)
 	if err != nil {
 		t.Fatal(err)
