+++ dsp/lutmath/trig_test.go
@@ -11,7 +11,7 @@ import (
 const testTolerance = 1e-6
 
 func TestSin(t *testing.T) {
-	want := math.Sin(39)
+	want := float32(math.Sin(39))
 	got := lutmath.Sin(39)
 	if !testutil.EqualWithinTolerance(want, got, testTolerance) {
 		t.Errorf("Sin(39) = %v, want %v", got, want)
@@ -19,7 +19,7 @@ func TestSin(t *testing.T) {
 }
 
 func TestCos(t *testing.T) {
-	want := math.Cos(39)
+	want := float32(math.Cos(39))
 	got := lutmath.Cos(39)
 	if !testutil.EqualWithinTolerance(want, got, testTolerance) {
 		t.Errorf("Cos(39) = %v, want %v", got, want)
@@ -27,7 +27,7 @@ func TestCos(t *testing.T) {
 }
 
 func TestTan(t *testing.T) {
-	want := math.Tan(39)
+	want := float32(math.Tan(39))
 	got := lutmath.Tan(39)
 	if !testutil.EqualWithinTolerance(want, got, testTolerance) {
 		t.Errorf("Tan(39) = %v, want %v", got, want)
